Fix everything-is-black-in-gingerbread styling prob introduced in 98c3025a
The styles are all defined by ActionBarSherlock, there's no need to
declare them here... and they sent the styling all very black on
pre-HoneyComb devices where ABS is triggered.

It would really *awesome* if our CI system could take screenshots with
every build, so we can quickly eyeball the appearance of the app to
compare the effect of changes.