Auto-update manifest version name & code using android-maven-plugin
This change turns on manifest-update with the android-maven-plugin:

http://maven-android-plugin-m2site.googlecode.com/svn/manifest-update-mojo.html#manifest

The version name & code are auto-updated to reflect the pom version, which
as usual is incremented by the maven release process (has the nice
side-effect of git-tagging the code).

Note that the manifest files will be auto-formatted using JAXP transformer
defaults.


Also simplifying proguard config by using the android-maven-plugin
proguard goal (which is somewhat stricter- if you have a method that is
*only* used in your integration test project, you need to explicitely
specify keeping it in the proguard.cfg)