Fix broken GitHubApiTest integration test
It's kinda interesting how the test was broken by this innocuous change:

https://github.com/github/android/commit/ec759a9

The change is that the 'getRepositories()' method is no longer used in
the main app (replaced by 'getOrgRepositories()'), and so is stripped by
proguard, but it's still used in the integration test. With recent
changes to the android-maven-plugin, the contributing library is only
processed by proguard once, so the explicit usage in the integration
tests isn't enough to get the 'getRepositories()' method kept - an
explicit proguard rule has to be added.