Add sync for users, orgs and their repos
This periodically fetches and stores that data for the user's account.

Note that it's currently hard-coded to only look at the user's 1st
account - while we only want the user to have a single github.com
registered, we'd like to be able to support an additional Enterprise
account(s) on the same device, so will need to revisit fore that.

Probably should revisit the behaviour here wrt battery use in light of:

http://developer.android.com/training/efficient-downloads/index.html