Remove Avatar sqlite storage as it's now on disk, post 49d5205
The default implementation of lazy-drawables storage is to cache to the
filesystem, which I think is appropriate for avatar blobs, as they are
not going to have db queries constructed on them.

Because our avatar file store is in the app's 'cache' folder we also get
a couple of nice things for free: the folder will be automatically
cleared by the operating system if disk space is tight, which is better than
us swelling a database file which can't be easily shrunk by the OS. The
android platform also gives the user the option to clear an apps cache
on the 'manage app' page - the user can use this as a brute-force
method to refresh their avatars.