Fix recently-viewed repos - 'Recent' labels now show correctly
This changes the implementation of RecentReposHelper to base it off a
single recently-used list, rather than one list per User/Org context.
I think this'll make it easier to add repos to the recently used-list
when spontaneously viewing repos without an explicit User/Org context,
and would make generating an overall list of most-recently viewed repos
easier if one was required.

The list of recently viewed repos help is now longer, at a maximum of 20
items, but only 5 'recently viewed' repos are displayed in each Org
context - the extra is to ensure that viewing 5 or more repos in one Org
context doesn't completely deplete the recent-repos in another.