A dummy file that will hopefully force installation of SDK 15
We need SDK 15 for compiling against ABS 4, but we don't want to *run*
the integration tests against it - we want to ensure our tests run
against the oldest version of Android we support, currently level 8.

The Janky build is configured to run against level 8, so that's what the
Android-Emulator-Plugin kindly installs for us - but we'd like it to
install level 15 as well. Hopefully this file will trigger that:

https://github.com/jenkinsci/android-emulator-plugin/blob/8a087d83a31975a008a9a90b2782bf871acd6ce8/src/main/java/hudson/plugins/android_emulator/builder/ProjectPrerequisitesInstaller.java#L83-102