add integration test for Org switcher and Org repo-list display
This just pages through the Orgs and checks that the repo list updates
to show at least one repo name known to be in the selected org. I've
added this test because I recently unwittingly broke the Org switcher
with d1ecbe8- the associated code is fragile and I'd like to make sure I
don't break it again! I'm a little concerned that the integration test
itself might be rather brittle but time will tell on that front...

This commit includes a tweak to cope with pre-ICS Spinner by ABS:
Instead of android.widget.Spinner, ABS uses
'com.actionbarsherlock.internal.widget.IcsSpinner' for the actionbar
dropdown on pre-ICS devices, so the integration test was failing to find
the dropdown on pre-ics devices.