Use precisely sized list to hold orgs. and default account
Previously an array resize would occur when the default
account was added to the list originally seeded with all
the orgs.

Also the returned list will not be created if either of
the API calls throws an exception where previously it would
be since the API calls were made after list creation.