Load pull request when issue load returns 410
Events in the dashboard that provide an issue but
are really pull requests can now be opened for
repositories that have issues disabled.

IssueStore now checks if an issue fetch failure is
due to a 410 indicating that issues are disabled and
the pull request API must be used for the issue model.