Handle case where first loader fails
Previously if the first loader throws an exception
then the empty view will be shown and subsequent calls
to setListShown will be ignored since listShown is
already set to true.

An additional check has been added when the shown value
is unchanging but the empty state may have changed so
a subsequent successful refresh will now show the list
and hide the empty view with no fade effect.