Show dialog when authenticator conflict exists
A SecurityException is raised when multiple apps
provide authenticator services for the same account
type id.  This causes the GitHub app to crash when a
another app has already registered an account using
this id (com.mobile).

The workaround is to catch the exception and show
a dialog describing the conflict and how it can
be resolved.

Closes issue #72