Delegate menu creation from activity to current fragment
Menu items would previously occasionally not be selectable
on Gingerbread and possibly other platforms when fragments
in a ViewPager were contributing menu items.  Now the
activity delegates directly to the fragments to create
the options menu and the fragments no longer call
setHasOptionsMenu which appears to remove the problem and
the menu items are now always selectable.

Closes issue #178